requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.68
construct-typing==0.5.6
cryptocompare==0.7.6
exceptiongroup==1.2.0
fake-useragent==1.5.1
filelock==3.15.1
filetype==1.2.0
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jito_searcher_client==0.1.3
jsonalias==0.1.1
lxml==5.2.2
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
nest-asyncio==1.6.0
ntscraper==0.3.13
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
pyotp==2.9.0
PySocks==1.7.1
pyTelegramBotAPI==4.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
regex==2024.5.15
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
snscrape==0.7.0.20230622
solana==0.30.1
solders==0.17.0
soupsieve==2.5
SQLAlchemy==2.0.29
telebot==0.0.5
Telethon==1.36.0
tqdm==4.66.4
twikit==1.7.5
types-cachetools==4.2.10
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websocket==0.2.1
websockets==11.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
